tast-tests: Fix user removal in ChallengeResponse
Fix the hwsec.ChallengeResponseAuthSession test to actually remove the
test user on completion. Also remove the no-op RemoveVault() calls from
the user creation function, since they didn't work (it's not possible to
remove an actively mounted user) and even if they did they'd break the
test (as the "defer" statements are executed right after the user
creation, before we had a chance to test the login).

BUG=b:240619207
TEST="tast run $DUT hwsec.ChallengeResponseAuthSession", SSH into DUT,
     run "ls /home/.shadow" and check there no subdirs there

Change-Id: I86dfcd4b745bf0edffb77e1ff5823906e375002e