lacros.AudioRecord: Promote to CQ
This informational test passed last 20+ builds except for VM
instability. According to the failure messages, the encountered failures
are also encountered in other tests, many of which are in CQ.

Use parameters to promote this test for boards with the lacros_stable
variation. For lacros_unstable boards such as VM, this test stays as
informational.

BUG=b:233504932
TEST=tast run ${DUT} lacros.AudioRecord

Change-Id: I9b7911ae83a2c56c65685d1299727184c6a41a84