crostini: increase timeout to launch Terminal
Launching Terminal after restarting Crostini may take longer than one
minute, which could fail the tests that include the process to restart.

Managed to reproduce the timeout failure 1 out of 10 iterations on an
eve device in the lab.

Need to increase the timeout before Restart.* and RestartIcon.* get
promoted to critical. Otherwise they will bring flakiness in CQ as the
previous RestartApp did.

BUG=None
TEST=tast run <DUT> crostini.*

Change-Id: I968c40adbadc57847ef5676f1585d1ba69a36599