Add MemoryCanaryPerf
Add a new memory test MemoryCanaryPerf. It continuously allocates memory
to either the host ChromeOS or one of the VMS until the canary, such as
a Chrome tab or an Android app, is killed because of the memory shortage.
Compared to the already existing multivm.Lifecycle, the test is expected
to be more understandable and less flaky because of its simplicity.

BUG=None
TEST=Run the following:
tast run -var multivm.MemoryCanaryPerf.iterations=2 localhost:2224 multivm.MemoryCanaryPerf.app_host
Confirm that the test finishes successfully.

Change-Id: Ic8437f55146ac41fabe0f8885804b0516c2482b0