tast-tests: Change framesender tests to not use DFS channels
We need to incur a 1 min penalty to use DFS channels. This wait time was
not enforced on Gale routers, but is on OpenWRT routers. We change the
following tests to not use DFS channels as that isn't the behavior we're
trying to test: CSALeaveChannel,DisconnectReason,FunctionalAfterCSA.

BUG=b:240153905
TEST=tast run
wifi.{CSALeaveChannel,DisconnectReason,FunctionalAfterCSA}.(*) $DUT on
openwrt router as well as gale

Change-Id: Ibf97ee254f46ca3974b29254841ec49a37dc1d92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3794441
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Joseph Pang <pjlee@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>