Stabilize photo taking in Camera Roll test
Taking a photo occasionally fails, due to either the /sdcard/DCIM/Camera
directory on existing on the phone, or crash dialogs for unrelated apps
obscuring the UI.

If /sdcard/DCIM/Camera does not exist, it will be created automatically
when a photo is taken, so we can just ignore it and proceed. To fix the
crash dialog issue, use an ADB command to prevent any crash dialogs from
showing in the first place.

BUG=b:240377733
TEST=tast run <dut> phonehub.DownloadRecentPhoto

Change-Id: I541afbe4e2c36eaf5cb1b1ad188491439dc5f943