WIP(extension) tast-tests: lacros variants to ui.BubbleLauncherAnimationPerf
This test needs Lacros variant and requires to measure performance
metrics with no browser window open in the first place.

For lacros-chrome this scenario could be possible only with the
"KeepAlive" feature that allows to get the browser instance without
opening a first window.
Also, it needs to install lacros apps from lacros extension, which is
currently blocked by crbug.com/1309565.

See go/lacros-tast-porting for the porting guidelines.

BUG=b:216242056
TEST=tast run ui.BubbleLauncherAnimationPerf*

Change-Id: Ie00f8fd9b86199515b689bcf92482f80d42ded30