WIP tast-tests: lacros variants to assistant.CardElementAnimationPerf
DO NOT SUBMIT: search_plus_a_lacros passed when it is run alone, but
failed with the error when it is run along with others:
"Failed to collect histograms: query results not shown for query Mount Everest"

This test needs Lacros variant and new fixtures for Lacros.

The scenario could be possible with the Lacros "KeepAlive" feature
that allows to get the browser instance without opening a first window.

See go/lacros-tast-porting for the porting guidelines.

BUG=b:219799321
TEST=test run -checktestdeps=false $DUT assistant.CardElementAnimationPerf*

Change-Id: I6a9dca5ec544ae05e5c58bea60627fa8b4bb92b4