tast-tests: lacros variants to dev.RemoteDesktop
This test needs Lacros enabled, open Web Store to install the remote
desktop app on browser.
For Lacros there might be a race condition that may cause the account
to be not added to the page first time it loads from profile.
To get around it, cws.InstallApp implements retry logic to reload the
page then check the account is added.

See go/lacros-tast-porting for the porting guidelines.

BUG=chromium:1322246, b:220799400
TEST=tast run -checktestdeps=false ${DUT} dev.RemoteDesktop*
TEST=tast run -checktestdeps=false -var=user=<user> -var=pass=<pass> ${DUT} dev.RemoteDesktop.lacros

Change-Id: I6110ac93d79de9255f8ee94938ec24ea6da476bf