camera: Lengthen the timeout for CCAUIPerf
For devices have both front and back cameras, the default timeout for
subtests are 90 seconds.

However, for each subtest, it reserves 10 seconds for cleanup, and have
up to 24 seconds to wait for CPU idle. And for each camera, it needs at
least 20 seconds to measure CPU usages and 5 seconds sleep to wait for
CPU idle. Therefore, 90 seconds seems to be a bit too strict and the
test might hit timeout sometimes.

This CL lengthen the default timeout for subtests to 120 seconds.

BUG=b:240889924
TEST=tast run {DUT} camera.CCAUIPerf

Change-Id: I4835cde377dde27e874929f8748eb5b094a19099