tast-tests: Several smartcard algs in CrossVersion
Specify multiple algorithms when testing the challenge-response (smart
card) login in the hwsec.CrossVersionLogin test.

This is closer to real-world use cases, since typical smart cards and
typical middleware support all RSA-SSA signature algorithms (SHA-1,
SHA-256, etc.). Also this will act as a regression test for the bug when
multiple specified algorithms confused the authentication stack in
cryptohome.

Note that this commit doesn't regenerate all cross-version data
snapshots. It's a lengthy process, and it makes sense to postpone it
until we have other changes that require regenerating the snapshots.

BUG=b:240619207,crbug:1346798
TEST=tast run hwsec.CrossVersionLogin.current

Change-Id: I32e91ef1c47de1c2bdd6a36041d3708cc09a2e34