windowarrangementcuj: Improve "normal" window move & resize
In the clamshell variants, when the window is in the "normal" state, the
move and resize will each be performed with a single drag instead of two
drags. Also, the test will wait for the expected window bounds before
ending each of these drags. This CL will fix behavior where the browser
window is supposed to be resized smaller with one drag and then back to
its original size with another drag, but sometimes the second drag
actually moves the window instead of resizing it, especially for Lacros.

BUG=b:238673213
TEST=tast run <DUT> ui.WindowArrangementCUJ.clamshell_mode
TEST=tast run <DUT> ui.WindowArrangementCUJ.lacros
TEST=tast unit tests passed.

Change-Id: I61a448f76c05c1042622263b3c2f282558ce0b1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3803688
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>