cellular: Ensure the modem FW gets flashed at least one
With the introduction of the flag `use_modems_fw_info`, the FW only gets
flashed if any of the FWs in the manifest is different. This usually
works fine for devices with multiple FWs, but on devices with a single
FW, the test never flashes anything, since the FW is always the same.

BUG=b:240985783
TEST=tast run tast_sasukenl668 cellular.ModemFWManifestInstallation
tast run tast_storo360 cellular.ModemFWManifestInstallation

Change-Id: Ib59b9c354ddbe92d0e95b70d689f649731c1e72f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3803692
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>