windowarrangementcuj: End overview mode to complete the loop
RunClamShell is called repeatedly, and each call will end overview mode
in preparation for the next call. This is now important because after
CL:3789488, the ash.CleanUpDesks function no longer ends overview. Also,
with this CL, overview mode will be ended with a real user workflow
whereas ash.CleanUpDesks ended overview with the autotest API. It is
considered preferable that CUJ tests represent real user workflows.

Note: This is not really a problem for RunTablet, because combineTabs
switches to clamshell mode thereby ending overview mode.

BUG=b:241014084
TEST=tast run <DUT> ui.WindowArrangementCUJ.clamshell_mode

Change-Id: Ib31262d7caaccad2b32aebe69f0bd52ed60b47da