tast-tests: remove linux FileWatcher at end of test
FilesApp keeps a cache of watcher handles [1].  When the 2nd test runs, it
is picking up the cached handle from test 1 where the container is now
destroyed, and it doesn't create new watcher. A fix for the test, is to
manually change dir at the end of the test to force the Linux watcher to
be removed.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/chromeos/extensions/file_manager/event_router.cc;l=753-776;drc=ddce2fc4e67fd4500d29cbe5f4993b3fb8e4e2ba

BUG=b:238600197
TEST=tast run $DUT crostini.FilesAppWatch.bullseye_stable crostini.FilesAppWatch.buster_stable

Change-Id: I4d614c59648367cfc1173048a739bf601ec430f4