fingerprint_test_images: Corrupt FW by increment value of byte
Replacing byte with random value (without any checking) introduces risk
that new value will be the same as old value. As a result the firmware
won't be corrupted. The same problem appears when we try to corrupt
firmware using some constant - it won't work when old value is the same
as constant.

Fix this problem by incrementing value of byte by 1, so new value will
be always different.

BUG=b:153736350
TEST=Run 'tast run <IP> firmware.FpROOnlyBootsValidRW' on hatch and
     nocturne.
TEST=TAST unit tests passed.

Change-Id: Ic959be40753515b715192a6b5148b805dbc58103