tast-tests: Replace PortID constants with lookup via PortType
This change removes all the Chameleond PortID constants in favor
of looking up the PortIDs which are of a certain PortType, the
human-readable name of a port as returned by GetConnectorType. New
fetch functions have been added to the Chameleond interface along with
constants for known PortType values to make this new port
identification process simple. Tests that used the older process have
been updated to use the new process.

This was necessary to account for later versions of Chameleon devices
which do not have a consistent set of PortID values that are used to
identify ports.

Bluetooth peers which use an older version of the Chameleond
interface return no results for GetConnectorType for some custom
bluetooth ports. In order to account for this, a backup PortID to
PortType map based on the original PortIDs is used if calls to
GetConnectorType fail. Thus, even older btpeers can still use the same
new method for looking up PortID values by PortType.

BUG=b:241132806
TEST=tast unit tests passed, including the new ones that were added.
Also passed assistant.OpenSettingsAudioChameleon using a Chameleon
with an audio board. Also confirmed that the new Fetch* functions work
with a btpeer using the static PortID to PortType map workaround.

Change-Id: I94a403a96455d7acc6022a80464c0ff11117312e