tast-tests: Relaunch Files App if Offline toggle is disabled
On startup, it takes some time before DriveFS initiates a native
messaging connection to Chrome. If the test opens Files App before this
connection is complete, the Offline toggle will be disabled. So,
relaunch Files App in this case to refresh the toggle's state.

Also fix an issue with the test not finding the "Enable Offline" button.
This button was moved to be part of a system notification rather than
part of the Files App.

BUG=chromium:1338095
TEST=tast run $DUT filemanager.DrivefsDssOffline

Change-Id: I0945789d6ddad11a666d51ceff94e1f87a499473
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3805252
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Austin Tankiang <austinct@chromium.org>