tast-tests: Add rollback system state files to TPM soft clear
The soft clear of the TPM does not carry out a full reset of the
TPM, but removes the system files based on a list. Rollback files
are also expected to be cleared as part of this process but they
are not and, as a result, the DUT is not left in the same state
as it was before the test. Add '/var/lib/oobe_config_restore'
and '/var/lib/oobe_config_save' to SystemStateFiles.

Include a clarification comment to ClearTPM for future reference.

BUG=b:240259124
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace
tast run -var=updateutil.tlwAddress="localhost:7151" DUT
autoupdate.EnterpriseRollbackPreviousVersion.rollback_1_version

Change-Id: I8767ff7bd4f97d66ee13b02c1b06453d0976a5d4