tast tests: fix rollback tests after auto re-enrollment
Do not wait for enrollment screen during tests. Auto-enrollment
is triggered automatically after rollback now. As a result of
this update, tests are failing when waiting for the enrollment
screen.

This check it is not necessary now, but it was not really necessary
before because getting the OOBE connection is enough to be able to
verify the networks have been preserved. Remove this check from
rollback tests.

Reorganise the test so the milestone is the first thing we retrieve
in case we need to carry out these kind of checks.

I have verified that tests that run in previous version still pass
after the removal.

BUG=b:234546174, b:236359370
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace
tast run -var=updateutil.tlwAddress="localhost:7151" DUT
autoupdate.EnterpriseRollbackPreviousVersion.rollback_1_version

Change-Id: I0ca26d2f0816a7671b50805c06bc66a76ed14dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3807590
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>