Adding tast tests for URL redirect
Added a bunch of URL tests to cover os://, chrome:// navigations for
successful, unsuccessful and blocked navigations to make sure that
the general redirection works.

This test does not test all the available shortcuts as there are
more than 200 tests otherwise to run, which take a considerable
amount of time + they are in constant flux anyways (urls being
added and getting removed).

It is worth mentioning that the navigation has to go through the
omnibox navigation and the tast tests were somehow bypassing the
path which leads to navigation errors for os:// URLs. As such
the URL has to be entered through the omnibox to make sure that
any navigation from Lacros to Ash is happening.

BUG=chromium:1289918
TEST=URLRedirect

Change-Id: I14a9a1df671022ab35c3a043527c39b598009d86