graphics: opencl-cts: remove daily run of a subset of the CTS
This daily run was supposed to help detect regression asap.
In practice it is not used at all, and clvk.api_tests is doing it in a
more efficient way.
As this daily run takes from 15 to 30min on each platform, remove it
to free some time.

BUG=b:241341344
TEST=./fast_build.sh -T

Change-Id: Ia1750bf8cfed17f0c065693d99de88b7ed968c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3809174
Reviewed-by: Seewai Fu <seewaifu@google.com>
Auto-Submit: Romaric Jodin <rjodin@chromium.org>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Romaric Jodin <rjodin@chromium.org>