printpreview: Select printers by actual printer name
In SelectPrinter(), find the printer by its actual name rather than
name + detail text. For the virtual printer, the detail text is just the
name of the printer again, but for real printers, that text is either
empty or something like "Out of ink". Our ability to select a printer
shouldn't be affected by whether it's out of ink, so use only the
printer's actual name.

BUG=b:195063930
TEST=CQ

Change-Id: If6f687080631db69c41aa9e9a81019be70378684
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3787089
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Bryan Cain <bryancain@chromium.org>
Commit-Queue: Bryan Cain <bryancain@chromium.org>