policy: Extend timeout for SSLErrorOverrideAllowed test
The lacros test is still failing, but the screenshot clearly shows that the button is visible and I can't reproduce it on my test device. I can reproduce it artificially if I add an extra short timeout to the waitUntilExists() method call. Therefore, I assume that on the bots the default timeout is simply not enough to allow the error page to load.
Extending the timeout should hopefully do the trick.

BUG=b:237062417, b:238605876

Change-Id: Ib27fc8134ea11d74129cd17a29297830f7fb6687