tast-tests: Don't use feature ForceShowContinueSection in tests
The feature flag "ForceShowContinueSection" populates the continue
section with fake files. For integration tests, we should use real
files.

Refactor a utility function to populate the continue section out of
the test launcher.ShowContinueSection.*. Use the new function in
launcher.HideContinueSectionClamshell and Tablet. The code in the
function is unchanged, except to convert s.Fatal to errors.Wrap,
and to return a cleanup function and list of filenames.

BUG=chromium:1349618
TEST=tast run <dut> launcher.*ContinueSection*

Change-Id: I609266aa60407b72ca74fc893da5c308a9005731