tast-tests: Add remote bluetooth fixtures with support for btpeers
Existing bluetooth tests are all local tests using a local test
fixture. While a local test can connect to a btpeer, it requires
the IP address of each btpeer to be provided in a tast parameter
as the DUT cannot resolve lab hostnames. Thus, local tests that
use btpeers cannot be scheduled.

This change adds new remote test fixtures that configure similar
bluetooth test environments like the existing local bluetooth
test fixtures, in addition to more fixtures that connect
to different amounts of btpeers. With these new btpeer fixtures,
remote bluetooth tests that use them do not require a tast var
to be specified to identify the btpeers' addresses, so tests that
use them may be scheduled to be run like any other bluetooth test.
To allow for local development, the btpeers tast var is kept as
an optional override for btpeer addresses.

The TestConnectToBTPeers test has been added as an example remote
test that uses ones of these new remote fixtures.

A new BTTestService has been added to allow for bluetooth tests
to access local test libraries not available to remote tests. The
initial version of this service only has methods required for the
test fixture, but as more remote bluetooth tests are created it is
expected to be expanded upon.

BUG=b:217633791
TEST=tast unit tests passed. Ran test TestConnectToBTPeers
successfully with two btpeers.

Change-Id: I39b9ef045312af42e7ab5445b3e2927f852ae380