iperf: Add ability to get server side logs from Iperf
When running Iperf UDP test, client side results are inaccurate as they only include the data that was sent and not what was actually received. Though it is possible to get the server side results from the client when using the iperf CLI, it is not possible via the cmw500 API. This CL adds the ability to fetch the most recent results from the Iperf server.

BUG=b:238343555
TEST=tast run -var=callbox=<CB> -var=callboxManager=<CBM> <DUT> cellular.Iperf*

Cq-Depend: chromium:3769602, chromium:3795621
Change-Id: If4613cebe56a6ef7f186a4f7ec76844195f98730