tast-test: Improve saved desk renaming consistency
Currently, ARC apps that are forced closed by the save for later
functionality will sometimes unexpectedly exit out of library view when
renaming the save and recall saved desk. This CL attempts to retry to rename the save and recall saved desk. If this second attempt fails, then there's an actual failure that needs to be caught.

BUG=b:241829597
TEST=tast run <DUT IP> wmp.DeskTemplates*

Change-Id: I97bab1a38dfefd959758a1d85d3cc66c3010d94c