tast-test: Exit out of unexpected popup
Currently, there is no way to disable the "no location access"
popup that unexpectedly appears on DUTs even when location
services are turned on. This popup also is not captured by the
UI DOM tree so the only way to exit it is through
keyboard inputs.

Work around by exiting the popup when it appears.

BUG=b:241829597
TEST=tast run <DUT IP> wmp.DeskTemplates*

Change-Id: I97bab1a38dfefd959758a1d85d3cc66c3010d94c