tast-tests: Fix for potential flakes in UI tests
This CL may prevent potential causes of test flakes
with ash.SetWindowStateAndWait.

Currently it checks only if the window animation is done, but
the problem is that there is a chance that a new window that is
starting up and hasn't become visible yet may be not waited enough
in this function and it returns early even before finishing changing
the window state on display.
This change fixes it by adding more window conditions to be waited.

BUG=b:225229931
TEST=tast run '("group:mainline" && "dep:chrome" && "name:wmp*" && !informational)'

Change-Id: Ic38efc07517b52c76473f2491c0b1278065b10ca