tast-tests: Fix for potential flakes in UI tests
This CL may prevent potential causes of test flakes
with ash.SetWindowStateAndWait and WaitWindowFinishAnimating.

Currently it checks if the window animation is simply done, but
the problem is that there is a chance that a new window that is
being opened and hasn't become visible yet.
The animation check makes sense only when the window is fully
in the visible state. Otherwise, WaitWindowFinishAnimating would
return early even before finishing changing the window state on
display.

BUG=b:225229931
TEST=tast run '("group:mainline" && "dep:chrome" && "name:wmp*" && !informational)'

Change-Id: Ic38efc07517b52c76473f2491c0b1278065b10ca