faft: Improve on relevant functions to open CCD
Add SetCCDCapability() for setting CCD capabilities to a
desired state, and modify GetCCDCapability() so that in
addition to returning the state, it will also return whether
a capability is accessible. Relevant modifications were made
in helper.go to accommodate for the above changes, and a few
steps rearranged. Also, allow some delay for DUT to change its
CCD level in case a reboot wasn't required.

BUG=b:240649924
TEST=manual on octopus.bobba, octopus.ampton, asurada.hayato

Change-Id: I6400a27bc10da9d4190e870477caf514537ec74c