arc.KernelCrash: Fix test by changing directory to look for crash files
With crrev.com/c/2333354, crash files are created in the daemon-store
crash directory when a user is logged in.
Currently, arc.KernelCrash Tast tests are failing because they look for
crash files in the legacy location.
Fix the tests by changing the directory to look for crash files to the
daemon-store one (following crrev.com/c/2333434).

BUG=b:220460507
TEST=tast run <DUT> arc.KernelCrash.*

Change-Id: I11e1dd9fc576ff652bb3e56ca4762472be576a28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3820441
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>