tast-tests: Run gofmt project
This change runs gofmt on the repository, cleaning up incorrectly
formatted files that likely only were merged to the tree due to
pre-submit checks being ignored.

A couple of these files had other pre-submit issues which
prevented submission after just running gofmt, so those issues
were resolved as well.

One of the fixes would be too complex to include in this small CL,
so b:242218149 has been created and a comment was added to address
this. This CL had to be submitted with presubmit checks disabled in
order to make the change to that file.

In addition, some files with a test declared that were corrected with
gofmt were failing the pre-submit check due to a missing LacrosStatus
setting, so a default option was set for those files.

BUG=None
TEST=tast unit tests passed; presubmit checks pass except for the one
mentioned above split to b:242218149

Change-Id: Ia44970266175916c55e3d082d5532a9168b9e7ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3828402
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>