tast-tests: Refactor bluetooth bluez D-Bus client
The bluetooth package is primarily just a set of files that builds
structs for managing D-Bus calls to bluez adapter and device objects.
While this method works for the minor usage that it has thus far, it
makes it difficult to organize things in the bluetooth package that
are not related to bluez or D-Bus.

This change moves these bluez D-Bus related structs to a new "bluez"
package under bluetooth. This package change is the reason why most
of the files in this commit were modified, as everywhere something
in package bluetooth that was moved to package bluez was used had
to be updated to use it from bluez rather than bluetooth. In addition,
some files with a test declared that needed this package change were
failing the pre-submit check due to a missing LacrosStatus setting,
so a default option was set for those files.

This change also refactors the bluez D-Bus structs to use DBusObject
to manage the D-Bus specific objects. DBusObject has also been updated
with new convenience functions that simplify property setting and
retrieval. As a result, the bluez D-Bus structs are much more simple
and the groundwork has been laid to add more structs for additional
bluez objects later easily.

Due to the refactoring of DBusObject, presubmit check failures caught
an issue with the exiting CheckNumInstalledProfiles function in
hermes.go using the testing state object outside the test. This has
been updated with a small fix as well.

BUG=b:217633791
TEST=tast unit tests passed; Ran test TestConnectToBTPeers
successfully with two btpeers; presubmit checks passing

Change-Id: I0a4972477d4a81991a98077ac31c02750b21c8a8