tast-tests: Add SimpleBTPeerPair test and necessary BTS service updates
Adds support for classic and LE bluetooth pairing of emulated btpeers
via DUT bluez dbus, and a test that uses the new support to test
pairing with such devices supported by btpeers.

The new remote test, bluetooth.SimpleBTPeerPair, uses test cases to
specify the device types being tested so that results can be tracked
on a per-device basis. New classic and LE bluetooth devices may be
added later if needed to this test, and other pairing tests may use
this as an example of how pairing tests may be done in Tast.

In the initial development there was a requirement for pin-based
pairing, so dbus logic was added to provide the pin to the bluetooth
adapter. This is no longer necessary as all these devices now default
to not require a pin, but this logic was kept and is still used to
allow for future devices that may require a pin to be used in the same
test.

BUG=b:217633791
TEST=tast unit tests passed; bluetooth.SimpleBTPeerPair.* tests all
passing consistently across multiple testbeds

Change-Id: I258124a78997c66abf9a02cdeb4f730ae451b269