tast-tests: syzkaller: Fix isLocal logic
Currently the code will load/save corp to GSC if the `isLocal` variable
is not empty. If -var=syzkaller.isLocal="true" is added as an argument,
the test will load/save to GSC.

Fix the check to see if isLocal is equal to "true". Set the default
value to "false" as well.

BUG=b:230347149
TEST=tast -verbose run -var=syzkaller.isLocal="true" $IP syzkaller.Wrapper.non_brya
TEST=tast -verbose run $IP syzkaller.Wrapper.non_brya

Change-Id: Icafc61f233df6e24a57475379fc69057e5a1f2a0