tast-tests: syzkaller: Fix isLocal logic
Currently the code will load/save corp to GCS if the `isLocal` variable
is not empty. If -var=syzkaller.isLocal="true" is added as an argument,
the test will load/save to GCS.

Fix the check to see if isLocal is equal to "true". Set the default
value to "false" as well.

BUG=b:230347149
TEST=tast -verbose run -var=syzkaller.isLocal="true" $IP syzkaller.Wrapper.non_brya
TEST=tast -verbose run $IP syzkaller.Wrapper.non_brya
TEST=tast unit test passed

Change-Id: Icafc61f233df6e24a57475379fc69057e5a1f2a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3826601
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Zubin Mithra <zsm@chromium.org>
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>