tast: Mark wifi.check_intel_fw_dump.go as unstable
Intel has made changes to the fw-dump-extractor.pl in core72, so the
tast parser version intel_fw_extractor.go need to be updated for the
test to be able to check the validity of the FW dump. Marking the test
as unstable until the script intel_fw_extractor.go is updated.

BUG=b:242055720
TEST=none

Change-Id: I495348b7bd8d16dc362f03a9ebb5d6628a3e36b1