tast-tests: Remove legacy launcher test variants, part 1
The new "productivity launcher" has been the default since M-104.
Remove the test variants that exercise the legacy "peeking"
launcher. Keep the string "productivity_launcher" in the
remaining variants so we can use tastboard/stainless to look at
long-term test history with a stable name.

This is preparation for deleting the legacy launcher code in
chromium.

BUG=chromium:1351225
TEST=tast run <dut> launcher.*

Change-Id: I8235f8d7ec72dd9350b7bf7d7b7f76a60f3cd0aa