tast-tests: Move wifi test fixture Reinit call back to Reset
The wifi test fixture's Reinit was originally called in the fixture
Reset step, but was moved to PreTest to always trigger OpenWrt reboots
before every test. This caused b:241836836, as it no longer continues
to run the test if Reinit, and therefore PreTest, fails.

This change moves the Reinit call back to Reset and also breaks down
DUT reinitialization and router reinitialization into two separate
sub-functions, ReinitDUT and ReinitRouter. With this, ReinitRouter
may be additionally called during NewTestFixture so that OpenWrt
routers are still rebooted before every test. This is also a better
solution than doing this in PreTest, as it's more of an environment
cleanup step rather than a step related to the test actions.

Fixture lifecycle timeouts have also been adjusted to reflect these
changes. The SetUp and Reboot timeouts have been increased to account
for the router rebooting and the PreTest timeout has been removed as
PreTest no longer preforms any actions.

BUG=b:242226273,b:241836836
TEST=tast unit tests pass; network.DiagSecureWifi.* passes with a
Gale router; wifi.SimpleConnect.80211a passes with an OpenWrt router
and reboot did still happen prior to running the test.

Change-Id: I0d93cae5382bcbe9114c95bb2a3fd8efefd1b1fe