firmware: Fix CorruptBothFWBodyAB() test case
This test very often did not pass because the operations involved in
writing to the firmware took a lot of time in skylab. The time in
which the test has to finish has been extended, and typos have been
additionally corrected. In addition, checking the mode the chromebook
is in after restoring the firmware has been disabled, because in
developer mode it caused, problems.

BUG=b:194907767
TEST=tast run -var servo=${SERVO_IP} ${DUT_IP}
firmware.CorruptBothFWBodyAB.normal_mode
TEST=tast unit test passed.

Change-Id: I1d1becfa5a6197dd50f1cea0ccd0de6b3b4f6b0e