firmware: Fix CorruptBothFWBodyAB() test case
This test very often did not pass because the operations involved in
writing to the firmware took a lot of time in skylab. The time in
which the test has to finish has been extended, and typos have been
additionally corrected. In addition, checking the mode the chromebook
is in after restoring the firmware has been disabled, because in
developer mode it caused, problems.

BUG=b:194907767
TEST=tast run -var servo=${SERVO_IP} ${DUT_IP}
firmware.CorruptBothFWBodyAB.normal_mode
TEST=tast unit test passed.

Change-Id: I1d1becfa5a6197dd50f1cea0ccd0de6b3b4f6b0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3827744
Reviewed-by: Kshitij Shah <tij@google.com>
Reviewed-by: Joseph Pang <pjlee@google.com>
Tested-by: Przemysław Filipek <pf@semihalf.com>
Reviewed-by: David Stevens <stevensd@chromium.org>
Auto-Submit: Przemysław Filipek <pf@semihalf.com>
Commit-Queue: Przemysław Filipek <pf@semihalf.com>