Stabilize arc.MediaScanPerf.sdcard_vm
The "sm unmount" command often fails for the emulated volume (SDCard
volume) when the MyFiles volume is mounted.
In order to stabilize the test, unmount the MyFiles volume before
unmounting the emulated volume.

Also, waitForARCVolumeUnmount for the emulated volume is often stuck
when the volume's state is immediately transitioned to "checking".
Fix this by regarding "checking" as an unmounted state.

BUG=b:242507052
TEST=arc.MediaScanPerf.*

Change-Id: I676f41e25357648ea973289700f74b9e4f312746