tast-tests: Add telemetry and info reporting test
This test ensures multiple data collected by the reporting team is being surfaced from the device to the server on the TELEMETRY_METRIC and INFO_METRIC destinations.

It also adds the ability for tests to login or to skip login using the
GAIAEnrollForReporting function.

Bug=b:243067525
TEST=tast -verbose run DUT policy.TelemetryInfoReporting*

Change-Id: I53185353fe5f1ab9d2b2494bc0e9df800e6256b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3829054
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Alberto Juarez <albertojuarez@google.com>
Commit-Queue: Alberto Juarez <albertojuarez@google.com>