tast-tests: Add ash.WaitForAppWindow
WaitForAppWindow finds app window by matching the app ID associated
and waits until it becomes visible in UI tree.
It would be useful to call this after ash.WaitForApp that checks shelf
behaviors for the app or apps.Launch in order to make sure the app
window is visible in the tree and testable for UI testing.

BUG=b:225229931
TEST=tast_run wmp.VirtualDesksShortcuts*

Change-Id: I8cb0c7a3acf180ad1912cd109d429bdda600e33a