tast-tests: Add ash.WaitForAppWindow
WaitForAppWindow finds app window by checking window title and waits
until it becomes visible in UI tree.
It would be useful to call this after ash.WaitForApp that is
a different check for certain shelf behaviors to be done for the app
ahead.

Minor change: WaitForAppsCondition that doesn't need to be exported is
removed.

BUG=b:225229931
TEST=tast_run wmp.VirtualDesksShortcuts*

Change-Id: I8cb0c7a3acf180ad1912cd109d429bdda600e33a