graphics: opencl-cts: update after uprev of clvk to ad6d1b71
remove get_kernel_arg_info_compatibility as it is as artifact from an
older version of the OpenCL-CTS. It does not exist anymore

divide, half_divide, samplerless tests, kernel_image_methods tests,
contractions and mem_host_flags are now passing

api_min_max_parameter_size is now failing because of an adding check
that where missing

rsqrt is now failing because of the fix of the division.

buffers and events tests are failing. Link with the github bug in
buganizer.

Thread dimensions are now always passing on brya/volteer/dedede

fdim and half_powr are always failing on brya/volteer/dedede

BUG=b:241773396, b:227133190, b:227134906, b:242503043, b:227133126, b:227132309
TEST=./fast_build.sh -T
TEST=tast run ${dut} graphics.OpenclCts.${test}

Change-Id: I326cda6a62c9b62bdd6e6f92c9a70e121f1d02df