media/video: Update tast tests to pass in codec for VAAPI tests.
The decode-test binary has been updated such that it now requires a
command line flag to specify which codec should be used for decoding.
This commit updates the tast-tests which execute the decode-test binary
to pass in the correct argument for each codec.

BUG=b:239719493
TEST=./fastbuild.sh -T
TEST=tast run $volteer video.PlatformDecoding.vaapi_av1
TEST=tast run $volteer video.PlatformDecoding.vaapi_vp9_0_group1_buf
TEST=tast run $volteer video.PlatformDecoding.vaapi_vp8_inter
TEST=tast run $volteer video.PlatformDecoding.vaapi_h265_baseline
TEST=tast run $volteer video.PlatformDecoding.vaapi_hevc_main

All tests are passing except video.PlatformDecoding.vaapi_hevc_main,
which is failing for a reason unrelated to this commit. However, the
tests do verify that it is receiving the new command line argument,
which is the behavior changed in this commit.

Change-Id: I5c3d3fa82d05ac5a385b4201dd50ca99b9e460d7