tast: Add physical keyboard typing latency tests.
Add a performance test for key event latency on physical keyboard.
Use the UMA histogram InputMethod.KeyEventLatency which measures
the time spent in the input method (excludes time spent in the app).

To start, we'll measure the latency of US English and Chinese Pinyin.
The input data needs to be improved as it is not representative of
the real world.

Measure the mean, % of key events that are < 5ms (called "fast"),
% of key events that are < 20ms (called "acceptable"). This allows
us to capture the overall shape with the mean, as well as capture
any outlier values.

We will later define objectives for each of these metrics.

BUG=b:241994527
TEST=tast run <dut> inputs.PhysicalKeyboardPerformance*

Change-Id: I12cd3ff211cac31ac56b24bc127cfa65e5eaaa86