network: add timeout for base network for routing_fallthrough tests
Currently in routing_fallthrough tests we only have timeout for
verifying the new-setup primary network but not for the existing base
network. However, adding a new default network also affects
configuration for the existing networks and a short amount is actually
needed to stablized. (Observable in b/242644942 when the proxied v6
address needed to be removed for base network to work.) This patch adds
the timeout for verification of the base network as well.

BUG=b:242644942
TEST=tast run $DUT network.Routing*

Change-Id: Iecf807e0a10942b66e83a9a82b83b18cccf40a4d